+++ releasenotes.md
@@ -1,4 +1,5 @@
 Whats New in v8.1
 --------------------------
+- Improve DB backup operations
 - Various bug fixes and improvements
 - For full details please refer to the CHANGELOG